requirements.txt
gurobipy
camelot-py
opencv-python
ghostscript
